+++ packages/generator-teams/src/app/manifestGeneration/ManifestGeneratorFactory.ts
@@ -8,6 +8,8 @@ import { ManifestGenerator as ManifestGenerator18 } from "./manifestGenerators/g
 import { ManifestGenerator as ManifestGenerator19 } from "./manifestGenerators/generator19/ManifestGenerator";
 import { ManifestGenerator as ManifestGenerator110 } from "./manifestGenerators/generator110/ManifestGenerator";
 import { ManifestGenerator as ManifestGenerator111 } from "./manifestGenerators/generator111/ManifestGenerator";
+import { ManifestGenerator as ManifestGenerator112 } from "./manifestGenerators/generator112/ManifestGenerator";
+import { ManifestGenerator as ManifestGenerator113 } from "./manifestGenerators/generator113/ManifestGenerator";
 import { ManifestGenerator as ManifestGeneratorDevPreview } from "./manifestGenerators/generatorDevPreview/ManifestGenerator";
 import { ManifestGenerator as ManifestGeneratorM365DevPreview } from "./manifestGenerators/generatorM365DevPreview/ManifestGenerator";
 import { BaseManifestGenerator } from "./BaseManifestGenerator";
@@ -40,15 +42,31 @@ export class ManifestGeneratorFactory {
             manifestVersion: ManifestVersions.v111,
             schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.11/MicrosoftTeams.schema.json",
             manifestValue: "1.11",
-            default: true,
+            default: false,
+            hide: true
+        },
+        {
+            manifestVersion: ManifestVersions.v112,
+            schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.12/MicrosoftTeams.schema.json",
+            manifestValue: "1.12",
+            default: false,
             hide: false
+        },
+        {
+            manifestVersion: ManifestVersions.v113,
+            schemaUrl: "https://developer.microsoft.com/en-us/json-schemas/teams/v1.13/MicrosoftTeams.schema.json",
+            manifestValue: "1.13",
+            default: true,
+            hide: false,
+            comment: "With support for extending Teams apps to other parts of the Microsoft 365 ecosystem"
         },
         {
             manifestVersion: ManifestVersions.devPreview,
             schemaUrl: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json",
             manifestValue: "devPreview",
             default: false,
-            hide: false
+            hide: false,
+            comment: "Developer preview"
         }
         ,
         {
@@ -56,7 +74,8 @@ export class ManifestGeneratorFactory {
             schemaUrl: "https://raw.githubusercontent.com/OfficeDev/microsoft-teams-app-schema/preview/DevPreview/MicrosoftTeams.schema.json",
             manifestValue: "m365DevPreview",
             default: false,
-            hide: false
+            hide: false,
+            comment: "Developer preview for extended Teams apps"
         }
     ]
 
@@ -84,6 +103,10 @@ export class ManifestGeneratorFactory {
             return new ManifestGenerator110();
         } else if (manifestVersion == ManifestVersions.v111) {
             return new ManifestGenerator111();
+        } else if (manifestVersion == ManifestVersions.v112) {
+            return new ManifestGenerator112();
+        } else if (manifestVersion == ManifestVersions.v113) {
+            return new ManifestGenerator113();
         } else if (manifestVersion == ManifestVersions.devPreview) {
             return new ManifestGeneratorDevPreview();
         } else if (manifestVersion == ManifestVersions.m365DevPreview) {